requirements2.txt
ansitable==0.11.4
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colored==2.3.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==3.0.3
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.9
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.54.0
identify==2.6.10
idna==3.10
importlib_metadata==8.6.1
inputs==0.5
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.7
machinevision-toolbox-python==1.0.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mvtb-data==1.5
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==1.24.2
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pgraph-python==0.6.3
pigpio==1.78
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.1
pre_commit==4.2.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
rpi_lgpio==0.6
rpi_ws281x==5.0.0
scipy==1.12.0
six==1.16.0
smbus2==0.4.2
spatialmath-python==1.1.14
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.4.0
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
zipp==3.21.0
